+++ src/commands/destroy.rs
@@ -2,7 +2,7 @@ use std::str;
 
 use yansi::Paint;
 
-use crate::{
+use crate::lib::{
     logger,
     project::{load_project, Project},
     resource_graph::{EvaluateResults, ResourceGraph},
