+++ src/lib/state/v1.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 use serde::{Deserialize, Serialize};
 use serde_yaml::{to_value, Mapping, Value};
 
-use crate::roblox_api::{ExperienceConfigurationModel, PlaceConfigurationModel};
+use super::super::roblox_api::{ExperienceConfigurationModel, PlaceConfigurationModel};
 
 use super::{
     legacy_resources::{Input, LegacyResource},
