+++ src/main.rs
@@ -2,14 +2,7 @@ extern crate cookie;
 
 mod cli;
 mod commands;
-mod config;
-mod logger;
-mod project;
-mod resource_graph;
-mod roblox_api;
-mod roblox_auth;
-mod roblox_resource_manager;
-mod state;
+mod lib;
 mod util;
 
 #[tokio::main]
